+++ src/main/java/lukfor/progress/styles/MinimalProgressBarStyle.java
@@ -41,7 +41,7 @@ public class MinimalProgressBarStyle implements IProgressBarStyle {
 		return width;
 	}
 
-	public String getLabel(int worked, int total) {
+	public String getLabel(long worked, long total) {
 
 		// TODO: style. raw, percentage, bytes
