+++ src/main/java/lukfor/progress/styles/ModernProgressBarStyle.java
@@ -41,7 +41,7 @@ public class ModernProgressBarStyle implements IProgressBarStyle {
 		return width;
 	}
 
-	public String getLabel(int worked, int total) {
+	public String getLabel(long worked, long total) {
 
 		// TODO: style. raw, percentage, bytes
